+++ cmd/gowdk/main.go
@@ -27,7 +27,7 @@ import (
 )
 
 const (
-	version    = "0.2.3"
+	version    = "0.2.5"
 	buildUsage = "usage: gowdk build [--config <file>] [--debug] [--ssr] [--allow-missing-backend] [--target <name>] [--module <name>] [--out <dir>] [--app <dir>] [--bin <file>] [--wasm <file>] [--backend-app <dir>] [--backend-bin <file>] [files...]"
 	initUsage  = "usage: gowdk init [--force] [--tests] [--template <site|minimal>] [dir]"
 )
@@ -194,6 +194,7 @@ bin/
 
 @page home
 @route "/"
+@guard public
 @css default page
 
 build {
@@ -272,6 +273,7 @@ bin/
 
 @page home
 @route "/"
+@guard public
 
 view {
   <main>